+++ foam/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ['app', 'Foam']
+__all__ = ['app', 'util', 'Foam']
 
 
 import functools as f
@@ -10,11 +10,10 @@ f.cached_property = compat.functools.cached_property
 f.singledispatchmethod = compat.functools.singledispatchmethod
 
 
-from . import app
+from . import app, util
 from .base.core import Foam
-from .util.function import license
 
 
 __doc__ = Foam.__doc__
-__license__ = license(full_text=False)
+__license__ = util.function.license(full_text=False)
 __version__ = Foam.__version__