+++ foam/util/object/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ['data', 'email', 'timer', 'version']
+
+
+from . import data, email, timer, version